refactor(vm): split init and debug helpers
- extracted VM construction and frame setup to VMInit.cpp
- moved breakpoint helpers to VMDebug.cpp and slimmed VM.cpp
- updated build to compile new translation units

tests: cmake --build build && ctest --test-dir build --output-on-failure

- [x] Built on macOS/Linux (Debug).
- [x] ctest passed.
- [ ] Added/updated tests.
- [x] Code formatted; no new warnings.
- [x] No spec/API deviation (or ADR attached).
- [ ] Docs updated if behavior/API changed.
- [x] Updated docs/comments per policy.
- [ ] Added/updated file headers.
